# Search: Modify query for better performance