… pkg/services/sqlstore/searchstore/search_test.go